Queue System working
Everything is accounted for: app pause/destroy (i want the queue to run
only when the app is open, otherwise, don't). Right now it only takes
care of checking whether internet is available or not, and uploading the
pendings. I need to make it save the pendings to disk before trying to
deal with them, but this code will be in the same class that deals with
the elastic search (and any kind of operation related to data
persistence and retrieval).